Masked out nails. Small adjustments.
