+++ common/src/main/java/io/github/mortuusars/envelope/Envelope.java
@@ -7,7 +7,8 @@ import io.github.mortuusars.envelope.advancements.critereon.MailDeliveredTrigger
 import io.github.mortuusars.envelope.advancements.predicate.ItemOccludingBlockPredicate;
 import io.github.mortuusars.envelope.advancements.predicate.ItemPackagePredicate;
 import io.github.mortuusars.envelope.command.argument.AddressArgument;
-import io.github.mortuusars.envelope.integration.wood_good.EveryCompatIntegration;
+import io.github.mortuusars.envelope.integration.Mods;
+import io.github.mortuusars.envelope.integration.every_compat.EveryCompatIntegration;
 import io.github.mortuusars.envelope.util.bugger.Bugger;
 import io.github.mortuusars.envelope.world.block.mailbox.MailboxBlock;
 import io.github.mortuusars.envelope.world.block.mailbox.MailboxBlockEntity;
@@ -101,7 +102,7 @@ public class Envelope {
         SoundEvents.init();
         ArgumentTypes.init();
 
-        if(Platform.isModLoaded("everycomp")) {
+        if (Mods.EVERY_COMPAT.isLoaded()) {
             EveryCompatIntegration.init();
         }
     }
