+++ common/src/main/java/io/github/mortuusars/envelope/integration/Mods.java
@@ -4,6 +4,7 @@ import io.github.mortuusars.envelope.Platform;
 
 public class Mods {
     public static final Mod SABLE = new Mod("sable");
+    public static final Mod EVERY_COMPAT = new Mod("everycomp");
 
     public record Mod(String id) {
         public boolean isLoaded() {
