+++ gradle.properties
@@ -5,7 +5,7 @@ enabled_platforms=fabric,neoforge
 mod_id=envelope
 archives_name=envelope
 mod_name=Envelope
-mod_version=0.7.4
+mod_version=0.7.5
 mod_release_type=release
 mod_license=GPLv3
 maven_group=io.github.mortuusars.envelope
